+++ ProgramFunctions/Update.fnc.php
@@ -515,7 +515,7 @@ function _update84()
  * Update to version 8.5
  *
  * 1. profile_exceptions table: Add Admin Student Payments Delete restriction.
- * 2. STAFF_EXCEPTIONS table: Add Admin Student Payments Delete restriction.
+ * 2. staff_exceptions table: Add Admin Student Payments Delete restriction.
  *
  * Local function
  *
@@ -543,7 +543,7 @@ function _update85()
 			WHERE modname='Student_Billing/StudentPayments.php&modfunc=remove');" );
 
 	/**
-	 * 2. STAFF_EXCEPTIONS table
+	 * 2. staff_exceptions table
 	 * Add Admin Student Payments Delete restriction.
 	 */
 	DBQuery( "INSERT INTO staff_exceptions