+++ functions/Search.fnc.php
@@ -308,7 +308,7 @@ function Search( $type, $extra = null )
 				WHERE (SELECT CAN_USE
 					FROM " . ( User( 'PROFILE_ID' ) ?
 						"profile_exceptions WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'" :
-						"STAFF_EXCEPTIONS WHERE USER_ID='" . User( 'STAFF_ID' ) . "'" ) . "
+						"staff_exceptions WHERE USER_ID='" . User( 'STAFF_ID' ) . "'" ) . "
 					AND MODNAME=CONCAT('Users/User.php&category_id=', sfc.ID)
 					LIMIT 1)='Y'
 				AND cf.CATEGORY_ID=sfc.ID
@@ -329,7 +329,7 @@ function Search( $type, $extra = null )
 				WHERE (SELECT CAN_USE
 					FROM " . ( User( 'PROFILE_ID' ) ?
 						"profile_exceptions WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'" :
-						"STAFF_EXCEPTIONS WHERE USER_ID='" . User( 'STAFF_ID' ) . "'") . "
+						"staff_exceptions WHERE USER_ID='" . User( 'STAFF_ID' ) . "'") . "
 					AND MODNAME='Users/User.php&category_id='||cf.CATEGORY_ID
 					LIMIT 1)='Y'
 				AND (SELECT VALUE
@@ -348,7 +348,7 @@ function Search( $type, $extra = null )
 				WHERE (SELECT CAN_USE
 					FROM " . ( User( 'PROFILE_ID' ) ?
 						"profile_exceptions WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'" :
-						"STAFF_EXCEPTIONS WHERE USER_ID='" . User( 'STAFF_ID' ) . "'") . "
+						"staff_exceptions WHERE USER_ID='" . User( 'STAFF_ID' ) . "'") . "
 					AND MODNAME='Students/Student.php&category_id='||sfc.ID)='Y'
 				AND cf.CATEGORY_ID=sfc.ID
 				AND NOT exists(SELECT ''
@@ -368,7 +368,7 @@ function Search( $type, $extra = null )
 				WHERE (SELECT CAN_USE
 					FROM " . ( User( 'PROFILE_ID' ) ?
 						"profile_exceptions WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'" :
-						"STAFF_EXCEPTIONS WHERE USER_ID='" . User( 'STAFF_ID' ) . "'") . "
+						"staff_exceptions WHERE USER_ID='" . User( 'STAFF_ID' ) . "'") . "
 					AND MODNAME='Students/Student.php&category_id='||cf.CATEGORY_ID)='Y'
 				AND (SELECT VALUE
 					FROM program_user_config