+++ modules/Students/Student.php
@@ -84,7 +84,7 @@ if ( User( 'PROFILE' ) !== 'admin' )
 	if ( User( 'PROFILE' ) !== 'student'
 		&& ! User( 'PROFILE_ID' ) )
 	{
-		$can_edit_from_where = " FROM STAFF_EXCEPTIONS WHERE USER_ID='" . User( 'STAFF_ID' ) . "'";
+		$can_edit_from_where = " FROM staff_exceptions WHERE USER_ID='" . User( 'STAFF_ID' ) . "'";
 	}
 
 	$can_edit_RET = DBGet( "SELECT MODNAME " . $can_edit_from_where .
@@ -662,7 +662,7 @@ if (  ( UserStudentID()
 	else
 	{
 		$can_use_sql = "SELECT MODNAME
-			FROM STAFF_EXCEPTIONS
+			FROM staff_exceptions
 			WHERE USER_ID='" . User( 'STAFF_ID' ) . "'
 			AND CAN_USE='Y'
 			AND MODNAME LIKE 'Students/Student.php%'";