+++ CHANGES.md
@@ -289,8 +289,8 @@ Changes in 8.5.1
 Changes in 8.5
 --------------
 - Fix SQL error duplicate key value violates unique constraint "food_service_menus_title" in Menus.php
-- SQL add PRIMARY KEY to STAFF_EXCEPTIONS table in rosariosis.sql
-- SQL profile_exceptions & STAFF_EXCEPTIONS tables: Add Admin Student Payments Delete restriction in Update.fnc.php & rosariosis.sql
+- SQL add PRIMARY KEY to staff_exceptions table in rosariosis.sql
+- SQL profile_exceptions & staff_exceptions tables: Add Admin Student Payments Delete restriction in Update.fnc.php & rosariosis.sql
 - Add Admin Student Payments Delete restriction in Profiles.php & Exceptions.php
 - Add Admin Student Payments Delete restriction in StudentPayments.php & Student_Billing/functions.inc.php
 - Fix SQL error numeric field overflow when entering percent > 100 in MassCreateAssignments.php